feat:move logic to servic
