+++ backend/http/auth/login.http
@@ -2,6 +2,6 @@ POST http://localhost:3001/auth/login HTTP/1.1
 Content-Type: application/json
 
 {
-  "email": "fidis18167@fgvod.com",
+  "email": "afterdeath@d.com",
   "password": "0000"
 }
