+++ backend/http/auth/register.http
@@ -2,8 +2,8 @@ POST http://localhost:3001/auth/register HTTP/1.1
 Content-Type: application/json
 
 {
-	"email": "fidis18167@fgvod.com",
+	"email": "sad@d.com",
     "password": "0000",
-    "name": "fidis Doe",
-    "organizationSlug": "hi fidis's Company"
+    "name": "dc Doe",
+    "organizationSlug": "hi dc's Company"
 }
